+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- use `strict` instead of `match` from `tcompare` for more verbose diffs
+
+### Fixed
+
+- correctly handle when shallow equality TAP output cannot be diffed
 
 ## [0.3.0] - 2022-03-23